+++ setup.cfg
@@ -1,0 +1,18 @@
+[metadata]
+name = QNS-2Q
+version = 0.0.1
+author = Muhammad Qasim
+author_email = mqasimk@gmail.com
+description = A package for simulating quantum non-local effects in a two-qubit system.
+long_description = file: README.md
+long_description_content_type = text/markdown
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
