+++ setup.py
@@ -1,16 +1,3 @@
 import setuptools
 
-setuptools.setup(
-    name="QNS-2Q",
-    version="0.0.1",
-    author="Muhammad Qasim",
-    author_email="mqasimk@gmail.com",
-    description="A package for simulating quantum non-local effects in a two-qubit system.",
-    packages=setuptools.find_packages(),
-    classifiers=[
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],
-    python_requires='>=3.6',
-)
+setuptools.setup()
