+++ src/IDopt.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import jaxopt
 import os
-from spectraIn import S_11, S_22, S_1_2, S_1212, S_1_12, S_2_12
+from .spectraIn import S_11, S_22, S_1_2, S_1212, S_1_12, S_2_12
 
 
 ########################################################################################################################
