+++ src/qnsplots.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import os
-from utils.spectraIn import S_11, S_22, S_1_2, S_1212, S_1_12, S_2_12
+from .spectraIn import S_11, S_22, S_1_2, S_1212, S_1_12, S_2_12
 
 
 specs1 = np.load('DraftRun_SPAM_hat/specs.npz')
