+++ src/spectraIn.py
