+++ src/trajectories.py
